presentai/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
google-genai==0.3.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-docx==1.1.0
PyPDF2==3.0.1
numpy==1.26.3
aiofiles==23.2.1
python-dotenv==1.0.0
